+++ packages/azure-kusto-ingest/src/index.ts
@@ -10,6 +10,12 @@ import managedStreamingIngestClient from "./managedStreamingIngestClient";
 import KustoIngestStatusQueues from "./status";
 
 import {
+    IngestionResult,
+    OperationStatus,
+    IngestionStatus,
+    IngestionStatusInTableDescription
+} from "./ingestionResult"
+import {
     DataFormat,
     IngestionMappingKind,
     ReportLevel,
@@ -78,6 +84,10 @@ export {
     ValidationPolicy,
     W3CLogFileMapping,
     dataFormatMappingKind,
+    IngestionResult,
+    OperationStatus,
+    IngestionStatus,
+    IngestionStatusInTableDescription
 };
 
 /**
@@ -106,7 +116,7 @@ export const IngestionPropertiesEnums = {
     SStreamColumnMapping,
     ConstantTransformation,
     FieldTransformation,
-    ColumnMapping,
+    ColumnMapping
 };
 
 export { IngestionPropertiesValidationError } from "./errors";