+++ packages/azure-kusto-ingest/src/ingestClient.browser.ts
@@ -5,10 +5,10 @@ import { KustoConnectionStringBuilder } from "azure-kusto-data";
 
 import { BlobDescriptor, generateBlobName, StreamDescriptor } from "./descriptors";
 import { FileDescriptor } from "./fileDescriptor.browser";
-import { QueueSendMessageResponse } from "@azure/storage-queue";
 
 import { IngestionPropertiesInput } from "./ingestionProperties";
 import { KustoIngestClientBase } from "./ingestClientBase";
+import { IngestionResult } from "./ingestionResult";
 
 export class KustoIngestClient extends KustoIngestClientBase {
     constructor(kcsb: string | KustoConnectionStringBuilder, defaultProps?: IngestionPropertiesInput) {
@@ -18,7 +18,7 @@ export class KustoIngestClient extends KustoIngestClientBase {
     /**
      * Use string for Node.js and Blob in browser
      */
-    async ingestFromFile(file: Blob | FileDescriptor, ingestionProperties?: IngestionPropertiesInput): Promise<QueueSendMessageResponse> {
+    async ingestFromFile(file: Blob | FileDescriptor, ingestionProperties?: IngestionPropertiesInput): Promise<IngestionResult> {
         this.ensureOpen();
         const descriptor = file instanceof FileDescriptor ? file : new FileDescriptor(file);
 
@@ -37,7 +37,7 @@ export class KustoIngestClient extends KustoIngestClientBase {
     /**
      * Use Readable for Node.js and ArrayBuffer in browser
      */
-    async ingestFromStream(stream: ArrayBuffer | StreamDescriptor, ingestionProperties?: IngestionPropertiesInput): Promise<QueueSendMessageResponse> {
+    async ingestFromStream(stream: ArrayBuffer | StreamDescriptor, ingestionProperties?: IngestionPropertiesInput): Promise<IngestionResult> {
         this.ensureOpen();
         const props = this._getMergedProps(ingestionProperties);
         const descriptor: StreamDescriptor = stream instanceof StreamDescriptor ? stream : new StreamDescriptor(stream);