+++ packages/azure-kusto-ingest/src/ingestClient.ts
@@ -6,11 +6,10 @@ import { KustoConnectionStringBuilder } from "azure-kusto-data";
 import { BlobDescriptor, generateBlobName, StreamDescriptor } from "./descriptors";
 import { FileDescriptor } from "./fileDescriptor";
 
-import { QueueSendMessageResponse } from "@azure/storage-queue";
-
 import { IngestionPropertiesInput } from "./ingestionProperties";
 import { KustoIngestClientBase } from "./ingestClientBase";
 import { Readable } from "stream";
+import { IngestionResult } from "./ingestionResult";
 
 export class KustoIngestClient extends KustoIngestClientBase {
     constructor(kcsb: string | KustoConnectionStringBuilder, defaultProps?: IngestionPropertiesInput) {
@@ -20,7 +19,7 @@ export class KustoIngestClient extends KustoIngestClientBase {
     /**
      * Use string in Node.JS and Blob in browser
      */
-    async ingestFromFile(file: FileDescriptor | string | Blob, ingestionProperties?: IngestionPropertiesInput): Promise<QueueSendMessageResponse> {
+    async ingestFromFile(file: FileDescriptor | string | Blob, ingestionProperties?: IngestionPropertiesInput): Promise<IngestionResult> {
         this.ensureOpen();
         const props = this._getMergedProps(ingestionProperties);
 
@@ -42,7 +41,7 @@ export class KustoIngestClient extends KustoIngestClientBase {
     async ingestFromStream(
         stream: StreamDescriptor | Readable | ArrayBuffer,
         ingestionProperties?: IngestionPropertiesInput
-    ): Promise<QueueSendMessageResponse> {
+    ): Promise<IngestionResult> {
         this.ensureOpen();
         const props = this._getMergedProps(ingestionProperties);
         const descriptor: StreamDescriptor = stream instanceof StreamDescriptor ? stream : new StreamDescriptor(stream);