+++ packages/azure-kusto-ingest/src/ingestionProperties.ts
@@ -249,6 +249,11 @@ export class IngestionProperties {
                 }
             }
         }
+
+        if (this.reportMethod !== ReportMethod.Queue && this.reportLevel === ReportLevel.FailuresOnly)
+        {
+            throw new IngestionPropertiesValidationError("ReportLevel.FailuresOnly is not supported with ReportMethod.Table");
+        }
     }
 
     merge(extraProps: IngestionPropertiesInput) {