+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unversioned
 
+### Changed
+
+- IngestClient returns Promise<IngestionResult> instead of Promise<QueueSendMessageResponse>
+
+### Added
+
+- Support table status reporting like explained here https://learn.microsoft.com/en-us/azure/data-explorer/kusto/api/netfx/kusto-ingest-client-status
+
 ## [5.2.3] - 2023-11-07
 
 ### Fixed